Try to make sim depend on button press
This doesn't entirely work, though, because x1 and x2 still depend on n and lambda, as before. This strategy just adds a new dependency (the action button). In other words, the action button WILL cause the simulation to run (hooray!) but so will changes to any of the other input parameters, which is bad because if the simulation is expensive, we want it to be clear that the simulation is run if---and only if---the simulation button gets clicked.