Try to debug a weird error
For this app, I created a slider to choose how much of the mtcars dataset the user wants to plot. The slider goes from 1 to 32. The weird part is that when you set the slider to 1, no points appear, and when you set the slider to 2, only one point seems to appear. I wonder if there's a bounds-checking error here?